Added Cleanup for Package and Software Masters
The new version code has been updated to Prepend the version so the
latest one is always the first result in the Versions node. This way the
version check only has to grab the first version node in the list
instead of sorting and hopefully getting the right one.
At the end of each run the Package and Software Masters are re-sorted so
that they are in software.Name order. This is mainly for readability.